Breaking change: one transaction per norm
A design decision in io.rkn/conformity to allow multiple transactions
per norm results in norms that cannot be conformed atomically. This
incurs significant complexity in how conformity checks that a norm is
conformed, and if not, how it is subsequently conformed.

io.rkn/conformity checks if a norm is conformed by comparing the
number of transaction datoms associated with the norm in the db,
against the number of transactions that the norm is attempting to
transact. If the counts differ, the norm is considered
uncomformed. When it tries to conform a norm, io.rkn/conformity takes
the index of the transaction in the transaction list, and checks if a
transaction datom is associated with that index in the db.

There are two problems with this approach:

First, the number and order of transactions in a norm may change and
potentially lead to data being duplicated. This could happen in a
number of scenarios: the user may have designed the norms without
taking into account the implementation details of how conformity
checks and conforms norms, or maybe the resolution of a partially
failed norm requires the number and order of transactions in a norm to
change.

The second problem is that conformity cannot not tell if a norm
function has been properly conformed without first evaluating the
function. This is problematic for expensive norm functions, for
example those that process large amounts of data, or pull in data from
external sources like s3. Ideally a norm function should not be
evaluated if it has already been conformed in the db.

This commit simplifies the conformity library to allow only a single
transaction per norm. This ensures that transacting norms is
atomic. It also makes norms trivial to validate: if there is any
transaction datom in the db associated with the norm id, the norm is
considered conformed. It also ensures that norm functions are not
evaluated if they have already been conformed.

Anything you could do before with multiple transactions per norm, you
should still be able to do with one or multiple norms. Because this is
a breaking change, conformity's :txes and :txes-fn attributes have
been changed to :tx and :tx-fn to make this explicit. Also, because
this is a significant breaking change, there is probably no point in
submitting a pull request.